requirements.txt
fastapi
uvicorn
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-requests==0.34b0
opentelemetry-exporter-otlp-proto-grpc
psutil
scikit-learn
joblib
requests
pandas
numpy
tensorflow
Flask
fastapi-cli
scipy
uvicorn
prometheus_client